Ensure that Travis always builds against HEAD.
We use submodules to pull in external dependencies, but in particular for
toxcore, we want to make sure both repos are built with the other's most recent
master. In toxcore, we do this by cloning hstox master. In hstox, we have
submodules that need to be updated. To avoid having to make frequent commits of
submodule updates, we simply check out the master branch and pull from it.